pyproject.toml
[project]
name = "gia-agentic-short-v2"
version = "0.1.0"
description = "Agentic systems with LangGraph and Anthropic Claude"
requires-python = ">=3.11,<3.14"
dependencies = [
    "langgraph>=0.2.0",
    "langchain>=0.3.0",
    "langchain-anthropic>=0.3.0",
    "langsmith>=0.2.0",
    "langchain-community>=0.3.0",
    "langchain-tavily>=0.1.0",
    "tavily-python>=0.5.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.27.0",
    "langgraph-checkpoint-sqlite>=3.0.1",
]

[project.optional-dependencies]
dev = [
    "jupyter>=1.0.0",
    "ipykernel>=6.0.0",
    "langgraph-cli>=0.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[dependency-groups]
dev = [
    "jupyter>=1.0.0",
    "ipykernel>=6.0.0",
    "langgraph-cli[inmem]>=0.1.0",
]
